Extract all the MDC and SwiftUI logic to a wrapper
